102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
 int main(void) {
     int n = 50;
     int i;
-    unsigned int fib[n];
+    unsigned long int fib[n];
 
     fib[0] = 1;
     fib[1] = 2;
@@ -16,7 +16,7 @@ int main(void) {
         fib[i] = fib[i - 1] + fib[i - 2];
     }
 	for (i = 0; i < n; i++) {
-        printf("%llu", fib[i]);
+        printf("%lu", fib[i]);
         if (i < n - 1) {
             printf(", ");
         }
